requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
nest-asyncio==1.5.6
playwright==1.29.1
pyee==9.0.4
python-telegram-bot==20.0
rfc3986==1.5.0
sniffio==1.3.0
TikTokApi==5.2.2
typing_extensions==4.4.0
urllib3==1.26.14
